Also synchronizes from a static template directory now
A directory initramfs.template is searched for from the update target
path up towards the root of the filesystem tree, and the first one
found is used.

The contents of this directory tree are then synchronized into the
update target path, overwriting any outdated or missing files there.

This is done before the synchronization as it has happened in older
revisions of the update script.