Avoid error message when no shared libraries are installed yet
In previous revisions, the update-initramfs script failed when there
were no old *.so* files to be deleted in the lib-directory.

However, this was always the case for a newly branched repository
project, because the shared libraries are not version controlled
and will only be put there by the updater script itself.