Add script for encrypted backups of tmfs files
The idea is to symlink the new script somwhere into a tree of
confidential files which are kept in a tmpfs because they shall never be
written to unencrypted storage, and encrypted storage has not been set
up yet.

The script then needs to be augmented with a configuration file which
specifies the base directory of the directory tree to be backed up, and
also other paths such as where the encrypted archive should be put and
which password file to use.

The script then checks whether any changes have been made to the
directory tree since the last backup, and refrehes the backup if there
are.

The script has also be tuned for being directly run as a cron job. This
allows automatic regular updates of the files in the tmpfs.