Add backup suppression to mk-backup
This is useful to avoid inconsistent backups while heavily re-organizing
the backup directory tree while a cron job might trigger a backup at any
moment in the background.

Of couse, it would also be possible to disable the cron job temporarily.
But why bother. Backup suppression is much easier. Just call the script
with option -d to suppress backups. Call it later again with option -e
to re-enable backups.

Also convert the documentation in the comments to real offline
documentation which can be shown with option -h.

Also add option -V for displaying the script version.